Move get_free_port() to utils
